Add support for opening archive files
Properly open given arguments instead of an hard-coded path.

Add support for archive files (ZIPs especially as per the DAISY format)
in a hackish way which currently requires gvfs-fuse and is fragile in
that a third party application can close the archive on us as it is
implemented through a mount (archive:// URI).

For some reason the player doesn't find the playback duration with
archive members, but works with the same uncompressed files.